+++ src/components/pages/WordbookPage.tsx
@@ -1,0 +1,9 @@
+const WordbookPage = () => {
+  return (
+    <div>
+      <h1>Wordbook Page</h1>
+    </div>
+  );
+};
+
+export default WordbookPage;
