+++ src/layouts/Layout.tsx
@@ -1,0 +1,5 @@
+const Layout = () => {
+  return <div></div>;
+};
+
+export default Layout;
